renamed router to routes
