+++ src/main.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import App from './App.vue'
-import router from './router'
+import router from './routes'
 import axios from 'axios'
 import store from "./store/store"
 import Embed from 'v-video-embed'
